pyproject.toml
[project]
name = "demo-python-api"
version = "0.1.0"
description = "FastAPI hello world following recommended backend structure"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "python-json-logger>=4.0.0",
  "python-jose>=3.5.0",
  "passlib>=1.7.4",
  "bcrypt==4.2.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.10.0",
  "email-validator>=2.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]

[tool.uv]
default-groups = ["dev"]


